partial implementation of signed struct lifetime fix.
