+++ pyrsia_node/src/node_manager/model.rs
@@ -1,4 +1,6 @@
 pub mod artifact;
 
 pub mod package;
+pub mod package_version;
+
 use crate::node_manager::*;
