+++ pyrsia_node/src/node_manager/model/package.rs
@@ -1,9 +1,14 @@
+extern crate pyrsia_client_lib;
 extern crate serde_json;
-use serde_json::{Map, Value};
-use strum_macros::{EnumIter, EnumString};
 
+use pyrsia_client_lib::signed::Signed;
+use serde_json::{Map, Value};
+
+use signed_struct::signed_struct;
+
+#[signed_struct]
 #[derive(Debug)]
-pub struct Package {
+pub struct Package<'a> {
     name: String,
     pkg_type: String,
     namespace_id: String,
@@ -15,25 +20,4 @@ pub struct Package {
     project_url: Option<String>,
     project_name: Option<String>,
     versions: Vec<u8>,
-}
-
-#[derive(Debug)]
-pub struct PackageVersion {
-    id: String,
-    version: String,
-    pkg_id: String,
-    license_text: Option<String>,
-    license_text_mimetype: Option<LicenseTextMimeType>,
-    license_url: Option<String>,
-    creation_time: Option<String>,
-    metadata: Map<String, Value>,
-    tags: Vec<String>,
-    description: Option<String>,
-}
-
-#[derive(EnumIter, Debug, PartialEq, EnumString)]
-pub enum LicenseTextMimeType {
-    Text,
-    Html,
-    Xml,
 }
